+++ packages/tiny-cli/src/typed.ts
@@ -2,11 +2,11 @@ import {
   difference,
   groupBy,
   pickBy,
-  range,
   tinyassert,
   zipMax,
 } from "@hiogawa/utils";
 import { parseRawArgsToUntyped } from "./untyped";
+import { DEFAULT_PROGRAM, ParseError, formatTable } from "./utils";
 
 //
 // ArgSchema
@@ -250,44 +250,3 @@ ${formatTable(optionsHelp)}
     parse,
   };
 }
-
-export class ParseError extends Error {
-  static wrapFn<T>(message: string, f: () => T): T {
-    try {
-      return f();
-    } catch (e) {
-      throw new ParseError(message, { cause: e });
-    }
-  }
-}
-
-//
-// help formatting
-//
-
-export function formatTable(rows: string[][]) {
-  return formatIndent(
-    padColumns(rows).map((row) => row.join(" ".repeat(4)).trimEnd()),
-    2
-  );
-}
-
-function padColumns(rows: string[][]): string[][] {
-  if (rows.length === 0) {
-    return rows;
-  }
-  const ncol = Math.max(...rows.map((row) => row.length));
-  const widths = range(ncol).map((c) =>
-    Math.max(...rows.map((row) => row[c]?.length ?? 0))
-  );
-  const newRows = rows.map((row) =>
-    row.map((v, i) => v.padEnd(widths[i], " "))
-  );
-  return newRows;
-}
-
-function formatIndent(ls: string[], n: number): string {
-  return ls.map((v) => " ".repeat(n) + v).join("\n");
-}
-
-export const DEFAULT_PROGRAM = "(cli)";
